requirements.txt
playwright>=1.42.0
python-dotenv>=1.0.1
rich>=13.7.1
